zevm: 64-bit stack pointer & indices (drop 32-bit address math)
Make EvmState::stackPointer, Regs::sp, stack_depth, ld_le/st_le and every
handler-local stack-index variable size_t instead of uint32_t. A 32-bit index
forces the rv64 codegen to zero-extend it (zext.w = slli/srli 0x20) and use addw
on every s.stack[sp+k] access; a 64-bit index folds straight into the 64-bit
address calculation. (evmone keeps a 64-bit stack-top pointer and pays none of
this.)

Measured (block 25343653, both ELFs gcc 14, ziskemu 1.0.0-alfa):
TOTAL 32.499B -> 31.173B (-4.08%), STEPS -3.48%. Narrows the gap to evmone from
+15.0% to +10.7%.

Correct: block 25343653 reproduces the chain hash; Prague EEST 2964/2965.

!!! TOOLCHAIN: build the RISC-V ELF with GCC 14 (xpack riscv-none-elf-gcc 14.x).
GCC 15.2.0 MISCOMPILES this widening at the guest's aggressive -O3 ZEG_GUEST_OPT
flags and yields a wrong state root (host g++ 13 + full EEST confirm the source
is correct, and the wrong hash differs between -O1 and -O3 — a codegen bug, not
UB). Pin the guest cross-compiler to GCC 14 until the GCC 15 bug is resolved.

+++ cpp-guest/src/zevm/evm_state.hpp
@@ -63,7 +63,7 @@ struct EvmState {
     // toward 0 (full). A push pre-decrements, a pop post-increments. Number of
     // live items == kStackLimit - stackPointer.
     U256           stack[kStackLimit];
-    uint32_t       stackPointer;       // kStackLimit == empty
+    size_t         stackPointer;       // kStackLimit == empty (64-bit: indexes the stack)
 
     // ----- memory -----
     // Handle into the static EVMMem manager (== this frame's call depth). The
